+++ glue/viewers/image/qt/tests/test_data_viewer.py
@@ -554,6 +554,32 @@ class TestImageViewer(object):
 
         assert self.viewer.layers[2].image_artist.get_visible()
 
+    def test_linking_and_enabling(self):
+
+        # Regression test for a bug that caused layers not not be correctly
+        # enabled/disabled.
+
+        self.viewer.add_data(self.image1)
+        self.viewer.add_data(self.catalog)
+        self.catalog.add_component([4, 5, 6], 'e')
+
+        self.data_collection.new_subset_group(subset_state=self.catalog.id['e'] > 4)
+
+        assert self.viewer.layers[0].enabled  # image
+        assert not self.viewer.layers[1].enabled  # scatter
+        assert not self.viewer.layers[2].enabled  # image subset
+        assert not self.viewer.layers[3].enabled  # scatter subset
+
+        link1 = LinkSame(self.catalog.id['c'], self.image1.world_component_ids[0])
+        link2 = LinkSame(self.catalog.id['d'], self.image1.world_component_ids[1])
+        self.data_collection.add_link(link1)
+        self.data_collection.add_link(link2)
+
+        assert self.viewer.layers[0].enabled  # image
+        assert self.viewer.layers[1].enabled  # scatter
+        assert not self.viewer.layers[2].enabled  # image subset
+        assert self.viewer.layers[3].enabled  # scatter subset
+
 
 class TestSessions(object):