+++ CHANGES.md
@@ -4,6 +4,9 @@ Full changelog
 v0.12.0 (unreleased)
 --------------------
 
+* Don't disable layer completely if it is not enabled, just disable checkbox.
+  Also show warnings instead of layer style editor. [#1451]
+
 * Generalize registry for data/subset actions to replace the former
   single_subset_action registry (which applied only to single subset selections).
   Layer actions can now be registered with the ``@layer_action`` decorator. [#1396]
@@ -14,6 +17,9 @@ v0.12.0 (unreleased)
 
 v0.11.2 (unreleased)
 --------------------
+
+* Fix a bug that caused the wrong layers to be enabled when mixing image and
+  scatter layers and setting up links. [#1451]
 
 * Remove 'sep' from menu on Linux. [#1394]